Aureus Analytics - Hackathon no. 3
Participated in a hackathon organized by Aureus Analytics and Uconn School of Business. The experience was enriching in terms of an insurance industry exposure and application of Analytics and Data Science in the field of ‘Insurance’ overall.

The product ‘CRUX’ is an innovative approach to improve an user experience by studying their behavior. 
We performed:

1. Data Exploration
2. Pattern Finding
3. Observations and Recommendation
4. Model tuning and Improvements

Tool: Power BI and Excel